generate activity move to controller
